Moved routes directory to root of server
