+++ server/routes/toolbox/Inventory.mjs
